+++ src/Raven.Server/Json/Parsing/IJsonParser.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Raven.Server.Json.Parsing
+{
+    public interface IJsonParser : IDisposable
+    {
+        void Read();
+        void ValidateFloat();
+    }
+}
