+++ src/Raven.Server/Json/Parsing/JsonParserToken.cs
@@ -1,0 +1,17 @@
+namespace Raven.Server.Json.Parsing
+{
+    public enum JsonParserToken
+    {
+        Null,
+        False,
+        True,
+        String,
+        Float,
+        Integer,
+        Separator,
+        StartObject,
+        StartArray,
+        EndArray,
+        EndObject
+    }
+}
