+++ src/Raven.Server/Json/RavenOperationContext.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using Raven.Server.Json.Parsing;
 using Sparrow;
 using Sparrow.Binary;
 using Voron.Impl;
